Fix a bug when resubscribing after the last unsubscription
If a process unsubscribed from a channel and it was the last process on that
channel, we sent UNSUBSCRIBE to Redis. However, we left an empty map of
subscribers to pids under that channel in "state.subscriptions". This meant that
resubscribing to that channel would not send the SUBSCRIBE message because
Redix.PubSub thought it was already subscribed (since the channel was in
"state.subscriptions", even if with no subscribers under it).

This commit fixes this problem.